requirements.txt
Django==2.0.3
django-cors-headers==2.2.0
djangorestframework==3.7.7
pkg-resources==0.0.0
pytz==2018.3
